backend/requirements.txt
# Echo Backend - Python Dependencies

# Core FastAPI framework and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Authentication and security
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Elasticsearch client for semantic search
elasticsearch==8.12.1
aiohttp==3.9.5

# NanoGPT API is OpenAI-compatible — called via httpx (already a dependency)

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP client for Ollama calls (pinned <0.28 — httpx 0.28 broke starlette TestClient)
httpx>=0.27.0,<0.28.0

# Sentence embeddings for semantic similarity search
sentence-transformers==3.0.1

# Database ORM for account storage
sqlalchemy==2.0.29
asyncpg==0.29.0
aiosqlite==0.20.0

# Email validation for Pydantic
email-validator==2.1.0.post1

# Testing dependencies
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==5.0.0
anyio==4.3.0
ruff==0.3.4
